refactor(model): flatten source structure
