+++ packages/model/src/events/ConnectionStateUpdated.ts
