+++ packages/model/src/events/VCAuthnEvent.ts
