+++ packages/model/src/index.ts
@@ -1,1 +1,3 @@
-export * from './models'
+export * from './messages'
+export * from './events'
+export * from './types'
