+++ packages/model/src/messages/BaseMessage.ts
