+++ packages/model/src/messages/CredentialIssuanceMessage.ts
