+++ packages/model/src/messages/InvitationMessage.ts
