+++ packages/model/src/messages/proofs/vc/VerifiableCredentialProofItemResult.ts
