+++ packages/model/src/messages/proofs/vc/VerifiableCredentialRequestedProofItem.ts
