+++ packages/model/src/messages/proofs/vc/VerifiableCredentialSubmittedProofItem.ts
