--- packages/model/src/models/index.ts
@@ -1,3 +1,0 @@
-export * from './messages'
-export * from './events'
-export * from './types'
